+++ include/Bounds.h
@@ -3,14 +3,14 @@
 
 struct Bounds
 {
-	double top;
-	double bottom;
-	double left;
-	double right;
+	float top;
+	float bottom;
+	float left;
+	float right;
 
 	Bounds() : top(0), bottom(0), left(0), right(0) {}
 
-	Bounds(double top, double bottom, double left, double right)
+	Bounds(float top, float bottom, float left, float right)
 		: top(top), bottom(bottom), left(left), right(right)
 	{
 	}
